+++ build.gradle
@@ -17,6 +17,6 @@ dependencies {
 
 test {
     useTestNG() {
-        suites "src/test/resources/Homework17.xml"
+        suites "src/test/resources/homeWork17.xml"
     }
 }
