+++ src/test/java/LoginTests.java
@@ -9,41 +9,19 @@ import org.testng.annotations.Test;
 import java.time.Duration;
 
 public class LoginTests extends BaseTest {
-
     @Test
     public void loginValidEmailPassword() throws InterruptedException {
-        //navigateToPage();
+        navigateToPage();
+        driver.get(url);
         provideEmail("elzat.nurmanbetova@testpro.io");
-        providePassword("Testpro12345!");
+        providePassword("kochkor123-Q!");
         clickLoginBtn();
         Thread.sleep(2000); // Sleep or pause for 2 seconds (adjust as needed)
 
-        // Expected Result
-
     }
 
-    @Test
-    public void loginInvalidEmailValidPassword() throws InterruptedException {
-        //navigateToPage();
-        provideEmail("demo@class.com");
-        providePassword("invalid");
-        clickLoginBtn();
-        Thread.sleep(2000); // Sleep or pause for 2 seconds (adjust as needed)
-        // Expected Result
-        Assert.assertEquals(driver.getCurrentUrl(), url); // https://qa.koel.app/
-    }
 
-    @Test
-    public void loginValidEmailEmptyPassword() throws InterruptedException {
-        //steps
-        //navigateToPage();
-        provideEmail("demo@class.com");
-        providePassword("");
-        clickLoginBtn();
-        Thread.sleep(2000); // Sleep or pause for 2 seconds (adjust as needed)
-        // Expected Result
-        Assert.assertEquals(driver.getCurrentUrl(), url); //https://qa.koel.app/
-        System.out.println("Just Testing console");
-    }
+
+
 
 }
