+++ src/test/resources/Homework17.xml
@@ -5,7 +5,7 @@
 
     <test name="Homework17" preserve-order="false">
         <classes>
-            <class name="Homework17"/>
+            <class name="homeWork17"/>
             <class name="BaseTest"/>
         </classes>
     </test>
